update demo script

+++ run.sh
@@ -5,6 +5,8 @@ echo "building parser..."
 cake build:parser
 echo "done\n\n\n"
 #bin/coffee --nodes examples/test.coffee
-rm examples/*.js
+if [-f examples/*.js]; then
+    rm examples/*.js
+fi
+
 bin/coffee -c examples/test.coffee
-cat examples/*.js
